Move AcquireEmailHelper.java so as to narrow down access to RC_SIGN_IN
