+++ auth/src/main/java/com/firebase/ui/auth/ui/email/SignInNoPasswordActivity.java
@@ -24,7 +24,6 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import com.firebase.ui.auth.R;
-import com.firebase.ui.auth.ui.AcquireEmailHelper;
 import com.firebase.ui.auth.ui.ActivityHelper;
 import com.firebase.ui.auth.ui.AppCompatBase;
 import com.firebase.ui.auth.ui.ExtraConstants;
